--- src/components/Docs/Policy.module.scss
@@ -1,34 +1,0 @@
-body .customPageWrapper {
-  background-color: $primary_background_color;
-  .react-pdf__Page__canvas {
-    margin: 0 auto;
-    background-color: red !important;
-  }
-}
-
-.Example {
-  margin: 1.6rem 0;
-}
-
-.Example__container__document {
-  width: 100%;
-
-  margin: 0 auto;
-}
-
-.Example__container__document .react_pdf__Document {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-}
-
-.Example__container__document .react_pdf__Page {
-  margin: 0 auto;
-
-  background-color: $primary_background_color !important;
-}
-
-.Example__container__document .react-pdf__Page__canvas {
-  margin: 0 auto;
-}
