+++ src/components/UserRoutes.tsx
@@ -1,7 +1,6 @@
 import { Suspense, lazy } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Loader from './Common/Loader/Loader';
-import Policy from './Docs/Policy';
 
 const HomePage = lazy(() => import('../pages/Home/Home'));
 const NotFoundPage = lazy(() => import('../pages/NotFound/NotFound'));
@@ -11,6 +10,7 @@ const CatalogPage = lazy(() => import('../pages/Catalog/Catalog'));
 const CartPage = lazy(() => import('../pages/Cart/Cart'));
 const ProductCardPage = lazy(() => import('../pages/ProductCard/ProductCard'));
 const UserAccountPage = lazy(() => import('../pages/UserAccount/UserAccount'));
+const Docs = lazy(() => import('../pages/Docs/Docs'));
 
 const UserRoutes = () => {
   return (
@@ -23,7 +23,8 @@ const UserRoutes = () => {
         <Route path="/cart" element={<CartPage />} />
         <Route path="/product/:id" element={<ProductCardPage />} />
         <Route path="/account" element={<UserAccountPage />} />
-        <Route path="/policy" element={<Policy />} />
+        <Route path="/policy" element={<Docs />} />
+        <Route path="/rules" element={<Docs />} />
         <Route path="*" element={<NotFoundPage />} />
       </Routes>
     </Suspense>
